Development

+++ bd/sql-queries/db.sql
@@ -39,8 +39,8 @@ CREATE TABLE Estacion_climatologica (
 CREATE TABLE Datos_Climatologicos(
     id_climatologicos int(10) NOT NULL AUTO_INCREMENT,
     fecha DATE,
-    precipitacion_mm FLOAT(3,1),
-    evaporacion_mm FLOAT(3,1),
+    precipitacion_mm FLOAT(5,1),
+    evaporacion_mm FLOAT(5,1),
     tmax FLOAT(3,1),
     tmin FLOAT(3,1),
     humedad_relativa FLOAT(4,2),
